requirements.txt
./wheels/finplanning_core-0.3.0-py3-none-any.whl
streamlit>=1.30,<2
plotly>=5.18,<6
pandas>=2.1,<3
streamlit-ace>=0.1.1,<1
pytest-playwright>=0.5,<1
